+++ sorter/src/main/java/sortpom/verify/ElementComparator.java
@@ -3,7 +3,7 @@ package sortpom.verify;
 import java.util.List;
 import org.dom4j.Element;
 import sortpom.util.XmlOrderedResult;
-
+import sortpom.util.XmlComparator;
 /**
  * @author bjorn
  * @since 2012-07-01
@@ -38,19 +38,6 @@ public class ElementComparator {
 
   private XmlOrderedResult isChildrenOrdered(
       String name, List<Element> originalElementChildren, List<Element> newElementChildren) {
-    var size = Math.min(originalElementChildren.size(), newElementChildren.size());
-    for (var i = 0; i < size; i++) {
-      var elementComparator =
-          new ElementComparator(originalElementChildren.get(i), newElementChildren.get(i));
-      var elementOrdered = elementComparator.isElementOrdered();
-      if (!elementOrdered.isOrdered()) {
-        return elementOrdered;
-      }
-    }
-    if (originalElementChildren.size() != newElementChildren.size()) {
-      return XmlOrderedResult.childElementDiffers(
-          name, originalElementChildren.size(), newElementChildren.size());
-    }
-    return XmlOrderedResult.ordered();
+     return XmlComparator.isChildrenOrdered(name, originalElementChildren, newElementChildren);
   }
 }
